requirements.txt
playwright==1.41.2
langchain==0.1.5
openai==1.12.0
python-dotenv==1.0.1
pydantic==2.6.1
aiohttp==3.9.3
beautifulsoup4==4.12.3
loguru==0.7.2